Render Stripe nudge on frontend for Premium Content block (#18245)
* Render Stripe nudge on frontend for Premium Content block

Renders a Stripe nudge on the frontend for admin users, if the
upgrade nudge is not already being displayed.

On wpcom sites the nudge directly links to the URL for connecting
to Stripe. On Jetpack sites it redirects to the post in the
editor, as the Stripe URL is not available from within the
render_callback in Jetpack.

* Use correct snake case

* Do not render block or nudge on Jetpack sites

The block will not be available on Jetpack sites so we no longer
need to handle this.